pyproject.toml
[project]
name = "sentinel"
version = "0.1.0"
description = "Passive RF drone detection system at 2.4 GHz"
requires-python = ">=3.10"
dependencies = [
    "numpy>=1.24",
    "scipy>=1.10",
    "matplotlib>=3.7",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21",
]
hardware = [
    "pyserial>=3.5",
]
# NOTE: UHD Python bindings (import uhd) come from the system package python3-uhd,
# not PyPI. The venv MUST be created with --system-site-packages to access them.
# On Ubuntu: sudo apt install python3-uhd

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "hardware: requires USRP hardware connected",
]
